+++ app/build.gradle
@@ -133,5 +133,6 @@ dependencies {
 
     implementation 'com.github.dhaval2404:imagepicker:2.1'    //для работы с картинками
     implementation 'com.yandex.android:maps.mobile:4.5.0-lite'//для ЯндексКарт
-    implementation "io.getstream:avatarview-coil:1.0.7" //avatars https://github.com/GetStream/avatarview-android
+    //https://github.com/GetStream/avatarview-android
+    implementation "io.getstream:avatarview-coil:1.0.7" //avatars
 }
